Moved tests to testPacages
