+++ src/test/java/tests/browse_languages/languages/JavaScriptLanguageTest.java
@@ -1,4 +1,4 @@
-package tests;
+package tests.browse_languages.languages;
 
 import base.BaseTest;
 import org.testng.Assert;
