+++ src/test/java/tests/guest_book/GuestBookV2Test.java
@@ -1,4 +1,4 @@
-package tests;
+package tests.guest_book;
 
 import base.BaseTest;
 import org.testng.Assert;
