+++ src/test/java/tests/guest_book/SignV2Test.java
@@ -1,4 +1,4 @@
-package tests;
+package tests.guest_book;
 
 import base.BaseTest;
 import org.openqa.selenium.Alert;
