+++ src/test/java/tests/top_lists/TopListEsotericTest.java
@@ -1,4 +1,4 @@
-package tests;
+package tests.top_lists;
 
 import base.BaseTest;
 import org.testng.Assert;
